DirectX行列型をXMFLOAT4X4に移行し構造を整理
DirectXの行列型を`XMMATRIX`から`XMFLOAT4X4`に変更し、HLSLとC++間のデータ整合性を向上させました。これに伴い、`Transform`クラスに`GetWorldMatrixFloat4x4`メソッドを追加し、行列のトランスポーズ処理を簡略化しました。

プロジェクト構成を整理し、不要なファイル（例: `Object.cpp`/`Object.hpp`）を削除、新たに`MainCamera`クラスを追加しました。これにより、シーン管理の仕組みが改善され、`SceneGame`にカメラを動的に追加する機能を実装しました。

また、`ModelLoader`や`SharedStruct`の変更により、`Mesh`型を`CB::Mesh`型に移行し、メッシュデータのロード処理を新しい構造体に対応させました。

その他、エラーチェックの追加やコードの可読性向上を目的としたリファクタリングを実施しました。

+++ RobotShootingGame/Source/Game/Scene/SceneBase.hpp
@@ -47,7 +47,7 @@ public:
 	inline T* AddGameObject(Layer layer = Layer::DEFAULT)
 	{
 		T* actor = new T();
-		m_Actors[layer].emplace_back(actor);
+		m_Actors[static_cast<size_t>(layer)].emplace_back(actor);
 		actor->Init();
 		return actor;
 	}
